modify descriptionList
